Added some kind of custom cache-aside pattern

+++ rev-rate-service/src/main/java/com/midnightsun/revrateservice/service/RatingService.java
@@ -4,7 +4,9 @@ import com.midnightsun.revrateservice.mapper.RatingMapper;
 import com.midnightsun.revrateservice.model.Rating;
 import com.midnightsun.revrateservice.repository.RatingRepository;
 import com.midnightsun.revrateservice.service.dto.RatingDTO;
+import com.midnightsun.revrateservice.service.redis.CacheService;
 import com.midnightsun.revrateservice.web.exception.HttpBadRequestException;
+import com.midnightsun.revrateservice.web.exception.HttpNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,10 +18,12 @@ public class RatingService {
 
     private final RatingRepository ratingRepository;
     private final RatingMapper ratingMapper;
+    private final CacheService cacheService;
 
-    public RatingService(RatingRepository ratingRepository, RatingMapper ratingMapper) {
+    public RatingService(RatingRepository ratingRepository, RatingMapper ratingMapper, CacheService cacheService) {
         this.ratingRepository = ratingRepository;
         this.ratingMapper = ratingMapper;
+        this.cacheService = cacheService;
     }
 
     public Page<RatingDTO> getAll(Pageable pageable) {
@@ -52,11 +56,22 @@ public class RatingService {
 
     private RatingDTO save(Rating rating) {
         final var savedRating = ratingRepository.save(rating);
+        cacheService.updateProductAverageRating(savedRating.getProductId());
         return ratingMapper.toDTO(savedRating);
     }
 
     public void delete(Long id) {
         log.debug("Request to delete RATING with ID: {}", id);
+
+        final var rating = ratingRepository.findById(id);
+
+        if (rating.isEmpty()) {
+            throw new HttpNotFoundException("Entity not found");
+        }
+
+        final var productId = rating.get().getProductId();
+
         ratingRepository.deleteById(id);
+        cacheService.updateProductAverageRating(productId);
     }
 }
