+++ rev-rate-service/src/main/java/com/midnightsun/revrateservice/service/ReviewService.java
@@ -4,7 +4,9 @@ import com.midnightsun.revrateservice.mapper.ReviewMapper;
 import com.midnightsun.revrateservice.model.Review;
 import com.midnightsun.revrateservice.repository.ReviewRepository;
 import com.midnightsun.revrateservice.service.dto.ReviewDTO;
+import com.midnightsun.revrateservice.service.redis.CacheService;
 import com.midnightsun.revrateservice.web.exception.HttpBadRequestException;
+import com.midnightsun.revrateservice.web.exception.HttpNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,10 +18,12 @@ public class ReviewService {
 
     private final ReviewRepository reviewRepository;
     private final ReviewMapper reviewMapper;
+    private final CacheService cacheService;
 
-    public ReviewService(ReviewRepository reviewRepository, ReviewMapper reviewMapper) {
+    public ReviewService(ReviewRepository reviewRepository, ReviewMapper reviewMapper, CacheService cacheService) {
         this.reviewRepository = reviewRepository;
         this.reviewMapper = reviewMapper;
+        this.cacheService = cacheService;
     }
 
     public Page<ReviewDTO> getAll(Pageable pageable) {
@@ -52,11 +56,22 @@ public class ReviewService {
 
     private ReviewDTO save(Review review) {
         final var savedReview = reviewRepository.save(review);
+        cacheService.updateProductReviews(review.getProductId());
         return reviewMapper.toDTO(savedReview);
     }
 
     public void delete(Long id) {
         log.debug("Request to delete REVIEW with ID: {}", id);
+
+        final var review = reviewRepository.findById(id);
+
+        if (review.isEmpty()) {
+            throw new HttpNotFoundException("Entity not found");
+        }
+
+        final var productId = review.get().getProductId();
+
         reviewRepository.deleteById(id);
+        cacheService.updateProductReviews(productId);
     }
 }
